Created CardUtils library
