+++ Dialogs/ConfirmOrderDialog.cs
@@ -1,12 +1,9 @@
-﻿using AdaptiveCards;
-using CoreBot.Store;
+﻿using CoreBot.Utilities;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Schema;
-using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System.Collections.Generic;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -51,7 +48,7 @@ namespace CoreBot.Dialogs
         {
             var userProfile = await _profileAccessor.GetAsync(stepContext.Context, () => new UserProfile());
 
-            var attachment = CreateCardFromOrder(userProfile);
+            var attachment = CardUtils.CreateCardFromOrder(userProfile);
 
             var opts = new PromptOptions
             {
@@ -108,58 +105,6 @@ namespace CoreBot.Dialogs
             await stepContext.Context.SendActivityAsync(MessageFactory.Text(whatElse), cancellationToken);
 
             return await stepContext.EndDialogAsync(null, cancellationToken);
-        }
-
-        private Attachment CreateCardFromOrder(UserProfile userProfile)
-        {
-            string[] paths = { ".", "Cards", "confirmOrderCard.json" };
-            var cardJson = File.ReadAllText(Path.Combine(paths));
-            var card = AdaptiveCard.FromJson(cardJson).Card;
-
-            //Ara hem convertit el JSON a un AdaptiveCard i editarem els fragments que ens interessen.
-
-            //Primer editem el FactSet (informació de l'usuari que sortirà a la fitxa).
-            var containerFact = (card.Body[1] as AdaptiveContainer);
-            var factSet = (containerFact.Items[1] as AdaptiveFactSet);
-            factSet.Facts.Add(new AdaptiveFact("Ordered by:", userProfile.Name));
-            factSet.Facts.Add(new AdaptiveFact("Company:", userProfile.Company));
-
-            //Ara editarem la informació que sortirà dels productes
-            var containerProducts = (card.Body[3] as AdaptiveContainer);
-
-            userProfile.ProductCart.Products.RemoveAll(item => item == null);
-
-            foreach (SingleOrder order in userProfile.ProductCart.Products)
-            {
-                AdaptiveColumnSet columns = new AdaptiveColumnSet();
-                AdaptiveColumn productColumn = new AdaptiveColumn();
-
-                AdaptiveTextBlock product = new AdaptiveTextBlock(order.Product);
-                product.Wrap = true;
-
-                productColumn.Width = "stretch";
-                productColumn.Items.Add(product);
-                columns.Columns.Add(productColumn);
-
-                AdaptiveColumn amountColumn = new AdaptiveColumn();
-
-                AdaptiveTextBlock amount = new AdaptiveTextBlock(order.AmountToString());
-                amount.Wrap = true;
-
-                amountColumn.Width = "auto";
-                amountColumn.Items.Add(amount);
-                columns.Columns.Add(amountColumn);
-
-                containerProducts.Items.Add(columns);
-            }
-
-            var attachment = new Attachment()
-            {
-                Content = JsonConvert.DeserializeObject(JsonConvert.SerializeObject(card)),
-                ContentType = "application/vnd.microsoft.card.adaptive"
-            };
-
-            return attachment;
         }
     }
 }
