+++ Permission/IPermissionObject.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace CoreBot.Interfaces
+namespace CoreBot.Permission
 {
     interface IPermissionObject
     {
